Add EGS_DynamicSource class
An EGS_DynamicSource allows the user to simulate dynamic (time-varying)
motion of any source during a single calculation.  Continuous motion
of the source is simulated between user-defined control points, where
each control point defines a set of incident polar coordinates and an
associated fractional monitor unit (mu) index.

If the source is an IAEA-format phase space file or a BEAM simualation
source, then, provided the source stores the value of mu for each
particle, the user has the option to use these when setting incident
source coordinates, thus synchronizing the motion of EGS_DynamicSource
with jaw and MLC opening coordinates in the upstream BEAM simulation.

For phase space and BEAM simulation sources, EGS_DynamicSource
essentially provides the same functionality as sources 20 and 21
in DOSXYZnrc, as documented by Lobo & Popescu, Phys. Med. Biol.,
55:4431-4443 (2010).

Note that implementation of this source required the implementation
of a new method, EGS_Float getMu(), in egs_base_source.  As
reimplemented in iaea_phsp_source and egs_beam_source, getMu returns
the value of mu associated with a particle.

See the source documentation for more info and an example input.